Added Centroid outline and corrected access modifiers in Point class,
Most of the Centroid class methods are complete but a few have been left
empty as to be created in compliance with literature on k-means process.

X and Y coordinate access modifiers in Point class changed to protected
so that Centroid subclass has access to change it without providing a
method to other packages.